separate openai s2t
